Rename application logger
