+++ libs/common/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/health/health.module.js';
-export * from './lib/logger/application.logger.js';
 export * from './lib/logger/logger-module.options.js';
+export * from './lib/logger/logger.js';
 export * from './lib/logger/logger.module.js';
