+++ libs/common/src/lib/health/health.controller.ts
@@ -1,14 +1,14 @@
 import { Controller, Get } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { HealthCheck, HealthCheckService } from '@nestjs/terminus';
-import { ApplicationLogger } from '../logger/application.logger.js';
+import { Logger } from '../logger/logger.js';
 
 @ApiTags('probes')
 @Controller('probes')
 export class HealthController {
   constructor(
     private readonly health: HealthCheckService,
-    private readonly logger: ApplicationLogger,
+    private readonly logger: Logger,
   ) {}
 
   @Get('liveness')
