+++ libs/common/src/lib/logger/logger.spec.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it, jest } from '@jest/globals';
-import { ApplicationLogger } from './application.logger.js';
+import { Logger } from './logger.js';
 
-describe('application logger', () => {
+describe('logger', () => {
   it('should log by level', () => {
     expect.assertions(1);
-    const logger = new ApplicationLogger({ enabled: false });
+    const logger = new Logger({ enabled: false });
     const loggerSpy = jest.spyOn(logger, 'logMessage');
     logger.verbose('Test', 'Context');
     logger.debug('Test', 'Context');
